+++ src/Storage/Storage.Management/Blob/NewAzureStorageContainer.cs
@@ -146,6 +146,9 @@ namespace Microsoft.Azure.Commands.Management.Storage
         [ValidateNotNull]
         public Hashtable Metadata { get; set; }
 
+        [Parameter(HelpMessage = "Enable object level immutability at the container level.", Mandatory = false)]
+        public SwitchParameter EnableImmutableStorageWithVersioning { get; set; }        
+
         public override void ExecuteCmdlet()
         {
             base.ExecuteCmdlet();
@@ -161,7 +164,7 @@ namespace Microsoft.Azure.Commands.Management.Storage
 
                 Dictionary<string, string> MetadataDictionary = CreateMetadataDictionary(Metadata, validate: true);
 
-                var contaienr =
+                var container =
                     this.StorageClient.BlobContainers.Create(
                             this.ResourceGroupName,
                             this.StorageAccountName,
@@ -170,9 +173,16 @@ namespace Microsoft.Azure.Commands.Management.Storage
                                 defaultEncryptionScope: this.DefaultEncryptionScope,
                                 denyEncryptionScopeOverride: this.preventEncryptionScopeOverride,
                                 publicAccess: (PublicAccess?)this.publicAccess,
-                                metadata: MetadataDictionary));
+                                metadata: MetadataDictionary,
+                                immutableStorageWithVersioning: this.EnableImmutableStorageWithVersioning.IsPresent ? new ImmutableStorageWithVersioning(true) : null));
 
-                WriteObject(new PSContainer(contaienr));
+                container =
+                    this.StorageClient.BlobContainers.Get(
+                            this.ResourceGroupName,
+                            this.StorageAccountName,
+                            this.Name);
+
+                WriteObject(new PSContainer(container));
             }
         }
     }